Changed more refs to RobotoMod and fixed up build.prop mess...
If it was kept how it was then in some cases build.prop would not install. Also, this is unclean and the OTA system couldn't identify the changed zip this way, because it extracts the uploaded ZIP and looks for /system/build.prop - but we didn't have one. This is all fixed now. PLEASE DO NOT CHANGE ANY BUILD.PROP STUFF WITHOUT PERMISSION FROM BENJAMINGWYNN!